Address breakpoint cache review feedback
